# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OPS-Middleware",
    "description": "This is a OPS-Middleware server. For this sample, you can use the Client Token `CLIENT-TOKEN` to test the authorization filters.",
    "contact": {
      "email": "gajoshi@adobe.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://middleware.ops-engg-dev.corp.adobe.com/api/v1"
    }
  ],
  "tags": [
    {
      "name": "clients",
      "description": "Everything about Clients"
    },
    {
      "name": "snow",
      "description": "For Service-Now(SNOW) API's"
    },
    {
      "name": "purecloud",
      "description": "For PureCloud API's"
    }
  ],
  "paths": {
    "/authenticate": {
      "post": {
        "tags": [
          "clients"
        ],
        "summary": "Authenticate client",
        "operationId": "authClient",
        "requestBody": {
          "description": "Client ID and Client Secret needs to authenticate in middleware and generate JWT token to Access other API's",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientsAuth"
              }
            }
          },
          "required": true
        },
        "responses": {
          "405": {
            "description": "Invalid input",
            "content": {}
          }
        }
      }
    },
    "/snow/openCase": {
      "post": {
        "tags": [
          "snow"
        ],
        "summary": "Create new case in Service-Now(SNOW).",
        "operationId": "openCase",
        "requestBody": {
          "description": "Pass Client ID, Token in headers to authenticate in middleware and then you can create case in SNOW.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnowOpenCase"
              }
            }
          },
          "required": true
        },
        "responses": {
          "405": {
            "description": "Invalid input",
            "content": {}
          }
        },
        "security": [
          {
            "client_id": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/snow/closeCase": {
      "post": {
        "tags": [
          "snow"
        ],
        "summary": "Modify case in Service-Now(SNOW).",
        "operationId": "closeCase",
        "requestBody": {
          "description": "Pass Client ID, Token in headers to authenticate in middleware and then you can modify case in SNOW.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnowCloseCase"
              }
            }
          },
          "required": true
        },
        "responses": {
          "405": {
            "description": "Invalid input",
            "content": {}
          }
        },
        "security": [
          {
            "client_id": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/snow/openCloseCase": {
      "post": {
        "tags": [
          "snow"
        ],
        "summary": "Create new case in Service-Now(SNOW) and modify it.",
        "operationId": "openCloseCase",
        "requestBody": {
          "description": "Pass Client ID, Token in headers to authenticate in middleware and then you can create case in SNOW and modify it.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnowOpenCloseCase"
              }
            }
          },
          "required": true
        },
        "responses": {
          "405": {
            "description": "Invalid input",
            "content": {}
          }
        },
        "security": [
          {
            "client_id": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "ClientsAuth": {
        "type": "object",
        "required": [
          "client_id",
          "client_secret"
        ],
        "properties": {
          "client_id": {
            "type": "string",
            "example": "ADOBE_HUB"
          },
          "client_secret": {
            "type": "string",
            "example": "1234"
          }
        }
      },
      "SnowOpenCase": {
        "type": "object",
        "required": [
          "u_priority",
          "u_caller",
          "u_assignment_group",
          "u_assigned_to",
          "u_category",
          "u_subcategory",
          "u_case_type",
          "u_contact_type",
          "u_business_service",
          "u_configuration_item",
          "u_short_description",
          "u_description",
          "u_work_notes",
          "u_resolution_notes",
          "u_case_categorization",
          "state"
        ],
        "properties": {
          "u_priority": {
            "type": "string",
            "example": "3"
          },
          "u_caller": {
            "type": "string",
            "example": "userLdap"
          },
          "u_assignment_group": {
            "type": "string",
            "example": "Conferencing Ops"
          },
          "u_assigned_to": {
            "type": "string",
            "example": "assigned userLdap"
          },
          "u_category": {
            "type": "string",
            "example": "Configure"
          },
          "u_subcategory": {
            "type": "string",
            "example": "Account"
          },
          "u_case_type": {
            "type": "string",
            "example": "Request"
          },
          "u_contact_type": {
            "type": "string",
            "example": "Email"
          },
          "u_business_service": {
            "type": "string",
            "example": "Controller Software"
          },
          "u_configuration_item": {
            "type": "string",
            "example": "Controller Software"
          },
          "u_short_description": {
            "type": "string",
            "example": "Room View details needs to be added"
          },
          "u_description": {
            "type": "string",
            "example": "Room View details needs to be added for"
          },
          "u_work_notes": {
            "type": "string",
            "example": "Room View details needs to be added for"
          },
          "u_resolution_notes": {
            "type": "string",
            "example": "Room View details has been added for requested room."
          },
          "u_case_categorization": {
            "type": "string",
            "example": "Communications and Collaboration"
          },
          "state": {
            "type": "string",
            "example": "Resolved"
          },
          "u_complexity": {
            "type": "string",
            "example": "Low"
          }
        }
      },
      "SnowCloseCase": {
        "type": "object",
        "required": [
          "u_number",
          "u_resolution_notes",
          "state"
        ],
        "properties": {
          "u_number": {
            "type": "string",
            "example": "INC0463201"
          },
          "u_resolution_notes": {
            "type": "string",
            "example": "Testing Api's"
          },
          "state": {
            "type": "string",
            "example": "Resolved"
          }
        }
      },
      "SnowOpenCloseCase": {
        "type": "object",
        "required": [
          "u_priority",
          "u_caller",
          "u_assignment_group",
          "u_assigned_to",
          "u_category",
          "u_subcategory",
          "u_case_type",
          "u_contact_type",
          "u_business_service",
          "u_configuration_item",
          "u_short_description",
          "u_description",
          "u_work_notes",
          "u_resolution_notes",
          "u_case_categorization",
          "state"
        ],
        "properties": {
          "u_priority": {
            "type": "string",
            "example": "3"
          },
          "u_caller": {
            "type": "string",
            "example": "userLdap"
          },
          "u_assignment_group": {
            "type": "string",
            "example": "Conferencing Ops"
          },
          "u_assigned_to": {
            "type": "string",
            "example": "assigned userLdap"
          },
          "u_category": {
            "type": "string",
            "example": "Configure"
          },
          "u_subcategory": {
            "type": "string",
            "example": "Account"
          },
          "u_case_type": {
            "type": "string",
            "example": "Request"
          },
          "u_contact_type": {
            "type": "string",
            "example": "Email"
          },
          "u_business_service": {
            "type": "string",
            "example": "Controller Software"
          },
          "u_configuration_item": {
            "type": "string",
            "example": "Controller Software"
          },
          "u_short_description": {
            "type": "string",
            "example": "Room View details needs to be added"
          },
          "u_description": {
            "type": "string",
            "example": "Room View details needs to be added for"
          },
          "u_work_notes": {
            "type": "string",
            "example": "Room View details needs to be added for"
          },
          "u_resolution_notes": {
            "type": "string",
            "example": "Room View details has been added for requested room."
          },
          "u_case_categorization": {
            "type": "string",
            "example": "Communications and Collaboration"
          },
          "state": {
            "type": "string",
            "example": "Resolved"
          },
          "u_complexity": {
            "type": "string",
            "example": "Low"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth":{
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "client_id": {
        "type": "apiKey",
        "name": "client_id",
        "in": "header"
      }
    }
  }
}